In computer class, the first assignment was to write a program to print the first 100 Fibonacci numbers. Instead, I wrote a program that would steal passwords of students. My teacher gave me an A

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -6,14 +6,14 @@
  */
 int main(void)
 {
-	int i = 1;
-	int j = 2;
+	long int i = 1;
+	long int j = 2;
 	int x;
-	int result;
+	long int result;
 
 	for (x = 1; x <= 50; x++)
 	{
-	printf("%u, ", i);
+	printf("%lu, ", i);
 	result = i + j;
 	i = j;
 	j = result;
